backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
google-genai==0.8.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.71.1
python-dotenv==1.0.1
numpy==2.1.3
scipy==1.14.1
pydantic==2.10.1
httpx==0.27.2
aiofiles==24.1.0
python-multipart==0.0.12
